rename utils.ts >> device-sizes.ts
