+++ components/Header/Header.styles.tsx
@@ -2,7 +2,7 @@ import styled from "@emotion/styled";
 import { Popover } from "@material-ui/core";
 import ArrowDropDownIcon from "@material-ui/icons/ArrowDropDown";
 import FavoriteBorderIcon from "@material-ui/icons/FavoriteBorder";
-import { pxIphone } from "../../utils";
+import { pxIphone } from "../../utilities/device-sizes";
 
 export const TopHeader = styled.div`
   padding: 10px 0 12px 0;
