+++ components/Home/Banner/Banner.styles.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { pxPC } from "../../../utils";
+import { pxPC } from "../../../utilities/device-sizes";
 import ButtonBase from "@material-ui/core/ButtonBase";
 import { XlargeTitle } from "../../../styles/BaseStyles";
 export const BannerContainer = styled.div`
