+++ components/Home/Home.styles.tsx
@@ -1,4 +1,4 @@
 import styled from "@emotion/styled";
-import { pxIphone, pxPC } from "../../utils";
+import { pxIphone, pxPC } from "../../utilities/device-sizes";
 
 export const Content = styled.div``;
