+++ components/Home/MobileLatest/MobileLatest.styles.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { pxIphone, pxPC } from "../../../utils";
+import { pxIphone, pxPC } from "../../../utilities/device-sizes";
 import Rating from "@material-ui/lab/Rating";
 import { DescText, PriceText } from "../../../styles/BaseStyles";
