+++ pages/about/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "@emotion/styled";
-import { pxPC } from "../../utils";
+import { pxPC } from "../../utilities/device-sizes";
 import { Layout } from "../../components";
 
 const Content = styled.div`
